refactor(starknet_l1_provider): extract transaction_manager into file
Only extract + visibility change
